增加次数管理

+++ hallServer/user/user_times.go
@@ -7,16 +7,18 @@ import (
 	"mj/hallServer/db"
 	"mj/hallServer/db/model"
 
+	"time"
+
 	"github.com/lovelly/leaf/log"
 )
 
 //表名字
 const (
-	day_time_table = "user_dat_times"
-	time_table     = "user_times"
+	day_time_table  = "user_dat_times"
+	week_time_table = "week_times"
+	time_table      = "user_times"
 )
 
-//非协程安全
 func (u *User) LoadTimes() {
 	//永久次数
 	times, err := model.UserTimesOp.QueryByMap(map[string]interface{}{
@@ -33,27 +35,54 @@ func (u *User) LoadTimes() {
 		"user_id": u.Id,
 	})
 
+	now := time.Now()
 	if derr == nil {
 		for _, v := range daytimes {
+			if !(v.CreateTime.Year() == now.Year() && v.CreateTime.Month() == now.Month() && v.CreateTime.Day() != now.Day()) { //创建时间不是今天
+				model.UserDayTimesOp.Delete(u.Id, v.KeyId)
+				continue
+			}
 			u.DayTimes[v.KeyId] = v.V
 		}
 	}
 
+	//每周次数
+	weektimes, werr := model.UserDayTimesOp.QueryByMap(map[string]interface{}{
+		"user_id": u.Id,
+	})
+
+	ny, nd := now.ISOWeek()
+	if werr == nil {
+		for _, v := range weektimes {
+			y, d := v.CreateTime.ISOWeek()
+			if !(ny == y && nd == d) { //创建时间不是一周
+				model.UserWeekTimesOp.Delete(u.Id, v.KeyId)
+				continue
+			}
+			u.WeekTimes[v.KeyId] = v.V
+		}
+	}
 }
 
 //永久次数
 func (u *User) GetTimes(k int) int64 {
+	u.RLock()
+	defer u.RUnlock()
 	return u.Times[k]
 }
 
 func (u *User) SetTimes(k int, v int64) {
+	u.Lock()
 	u.Times[k] = v
-	updateTimes(day_time_table, u.Id, k, v)
+	u.Unlock()
+	updateTimes(time_table, u.Id, k, v)
 }
 
 func (u *User) IncreaseTimes(k int, addv int64) {
+	u.Lock()
 	u.Times[k] += addv
-	updateTimes(day_time_table, u.Id, k, u.Times[k])
+	u.Unlock()
+	updateTimes(time_table, u.Id, k, u.Times[k])
 }
 
 func (u *User) GetTimrsAll() (data map[int]int64) {
@@ -66,16 +95,22 @@ func (u *User) GetTimrsAll() (data map[int]int64) {
 
 //每日次数
 func (u *User) GetDayTimes(k int) int64 {
+	u.RLock()
+	defer u.RUnlock()
 	return u.DayTimes[k]
 }
 
 func (u *User) SetDayTimes(k int, v int64) {
+	u.Lock()
 	u.DayTimes[k] = v
+	u.Unlock()
 	updateTimes(day_time_table, u.Id, k, v)
 }
 
 func (u *User) IncreaseDayTimes(k int, addv int64) {
+	u.Lock()
 	u.DayTimes[k] += addv
+	u.Unlock()
 	updateTimes(day_time_table, u.Id, k, u.DayTimes[k])
 }
 
@@ -89,17 +124,23 @@ func (u *User) GetDayTimrsAll() (data map[int]int64) {
 
 //每周次数
 func (u *User) GetWeekTimes(k int) int64 {
+	u.RLock()
+	defer u.RUnlock()
 	return u.WeekTimes[k]
 }
 
 func (u *User) SetWeekTimes(k int, v int64) {
+	u.Lock()
 	u.WeekTimes[k] = v
-	updateTimes(day_time_table, u.Id, k, v)
+	u.Unlock()
+	updateTimes(week_time_table, u.Id, k, v)
 }
 
 func (u *User) IncreaseWeekTimes(k int, addv int64) {
+	u.Lock()
 	u.WeekTimes[k] += addv
-	updateTimes(day_time_table, u.Id, k, u.WeekTimes[k])
+	u.Unlock()
+	updateTimes(week_time_table, u.Id, k, u.WeekTimes[k])
 }
 
 func (u *User) GetWeekTimesAll() (data map[int]int64) {
@@ -144,6 +185,22 @@ func (u *User) IncreaseTimesByType(id int, v int64, types int) {
 	}
 }
 
+//////////////////////////////////////
+
+func (u *User) ClearDayTimes() {
+	u.Lock()
+	u.DayTimes = make(map[int]int64)
+	u.Unlock()
+	ClearTimes(day_time_table, u.Id)
+}
+
+func (u *User) ClearWeekTimes() {
+	u.Lock()
+	u.WeekTimes = make(map[int]int64)
+	u.Unlock()
+	ClearTimes(week_time_table, u.Id)
+}
+
 func updateTimes(table_name string, uid int, k int, v int64) bool {
 	sql := fmt.Sprintf("insert into %s values(%d,%s,%d) on duplicate key update v=%d;", table_name, uid, k, v, v)
 	_, err := db.DB.Exec(sql)
@@ -154,8 +211,17 @@ func updateTimes(table_name string, uid int, k int, v int64) bool {
 	return true
 }
 
-func ClearTimes(table_name string) {
-	sql := fmt.Sprintf("delete from %s")
+func ClearTimes(table_name string, id int) {
+	sql := fmt.Sprintf("delete from %s where user_id=%d;", id)
+	_, err := db.DB.Exec(sql)
+	if err != nil {
+		log.Error("at updateTimes error:%s", err.Error())
+		return
+	}
+}
+
+func ClearTimesByKeys(table_name string, Uid, key int) {
+	sql := fmt.Sprintf("delete from %s where user_id=%d and key_id=%d;", Uid, key)
 	_, err := db.DB.Exec(sql)
 	if err != nil {
 		log.Error("at updateTimes error:%s", err.Error())
